pyproject.toml
[project]
name = "watermarker-tool"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.3.0",
    "pymupdf>=1.26.4",
]
 
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["watermark"]

[project.scripts]
watermark = "watermark.cli:main"
